+++ app/src/main/java/woowacourse/movie/moviereservation/MovieReservationPresenter.kt
@@ -2,6 +2,7 @@ package woowacourse.movie.moviereservation
 
 import android.util.Log
 import woowacourse.movie.model.HeadCount
+import woowacourse.movie.moviereservation.uimodel.HeadCountUiModel
 import woowacourse.movie.repository.MovieRepository
 import java.time.LocalDate
 import java.time.LocalDateTime
