+++ app/src/main/java/woowacourse/movie/moviereservation/uimodel/ScreeningDateTimeUiModel.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.moviereservation
+package woowacourse.movie.moviereservation.uimodel
 
 data class ScreeningDateTimeUiModel(
     val date: String,
